Updated vortex tracking methods to search for regions of vortex points
The regions of vortex points constitute a "vortex" in the physical sense. The primary data type used for this implementation is the map which allows for searches on the global index. The next step is to perform the synchronization of the tags and reorder them using a global tagging convention.